requirements.txt
# Python 3.10+
python-dotenv==1.2.2
google-genai==1.69.0
mcp==1.26.0

# Subprocess deps for @nitansde/mcp (Cloudflare bypass). Pass this venv's python via:
# MCP_SERVER_ARGS=["-y","@nitansde/mcp@latest","--python_path","/abs/path/to/.venv/bin/python"]
cloudscraper==1.2.71
curl-cffi==0.14.0
brotli==1.2.0
